refactored BuildingStore and ResourceStore into MobX directory
