+++ src/Screens/TrackerScreen.js
@@ -15,8 +15,8 @@ import {
   Ionicons,
   MaterialIcons,
 } from "@expo/vector-icons";
-import resourceStore from "../Helpers/ResourceStore";
-import buildingStore from "../Helpers/BuildingStore";
+import resourceStore from "../Helpers/MobX/ResourceStore";
+import buildingStore from "../Helpers/MobX/BuildingStore";
 import DraggableFlatList from "react-native-draggable-flatlist";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import { observer } from "mobx-react";
