+++ src/Screens/VotingScreen.js
@@ -14,11 +14,17 @@ import faunaStore from "../Helpers/FaunaStore";
 import { observer } from "mobx-react";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { useNavigation } from "@react-navigation/native";
-import Constants from "expo-constants";
+import featureSuggestionsStore from "../Helpers/MobX/featureSuggestionsStore";
+import {
+  getSuggestions,
+  voteOnSuggestion,
+} from "../Helpers/Firebase/firestoreDriver";
 
 const VotingScreen = observer(() => {
   // mobX
-  const { features, loadFeatures, saveVote } = faunaStore;
+  // const { features, loadFeatures, saveVote } = faunaStore;
+  const { suggestions, loadSuggestions, voteOnSuggestion } =
+    featureSuggestionsStore;
 
   // States
   const [voted, setVoted] = useState(false);
@@ -97,7 +103,7 @@ const VotingScreen = observer(() => {
     try {
       await AsyncStorage.setItem("lastVoteTimestamp", Date.now().toString());
       await AsyncStorage.setItem("lastVotedFor", featureVote);
-      saveVote(featureVote).then(() => console.log("Vote saved"));
+      voteOnSuggestion(featureVote).then(() => console.log("Vote saved"));
       setVoted(true);
 
       console.log("Vote submitted successfully");
@@ -106,16 +112,30 @@ const VotingScreen = observer(() => {
     }
   };
 
+  // useEffect(() => {
+  //   setLoading(true);
+  //
+  //   if (!features)
+  //     loadFeatures()
+  //       .then(() => {
+  //         console.log("Loaded features");
+  //         setLoading(false);
+  //       })
+  //       .catch((error) => console.log(error));
+  //
+  //   canVoteToday()
+  //     .then((res) => setVoted(!res))
+  //     .catch((error) => console.log(error));
+  //
+  //   setTimeout(() => {
+  //     setLoading(false);
+  //   }, 1000);
+  // }, [features, loadFeatures]);
+
   useEffect(() => {
     setLoading(true);
 
-    if (!features)
-      loadFeatures()
-        .then(() => {
-          console.log("Loaded features");
-          setLoading(false);
-        })
-        .catch((error) => console.log(error));
+    loadSuggestions().then(() => console.log("Loaded suggestions"));
 
     canVoteToday()
       .then((res) => setVoted(!res))
@@ -124,7 +144,7 @@ const VotingScreen = observer(() => {
     setTimeout(() => {
       setLoading(false);
     }, 1000);
-  }, [features, loadFeatures]);
+  }, []);
 
   return (
     <SafeAreaView style={styles.container}>
@@ -146,7 +166,7 @@ const VotingScreen = observer(() => {
       {/*  simply increments a tally.*/}
       {/*</Text>*/}
 
-      {!features ? null : (
+      {!suggestions ? null : (
         <Text
           style={{
             margin: 20,
@@ -161,19 +181,17 @@ const VotingScreen = observer(() => {
           }}
         >
           There are currently{" "}
-          <Text style={{ fontWeight: "800" }}>{features.length}</Text> features
-          to vote from. Scroll through the list below and read the description
-          to get a better idea what each feature is about:
+          <Text style={{ fontWeight: "800" }}>{suggestions.length}</Text>{" "}
+          features to vote from. Scroll through the list below and read the
+          description to get a better idea what each feature is about:
         </Text>
       )}
 
       <ScrollView>
         {loading ? <ActivityIndicator size="large" color="#d35322" /> : null}
-        {!features
+        {!suggestions && !loading
           ? null
-          : features.map((item, index) => {
-              const feature = item.data;
-
+          : suggestions.map((feature, index) => {
               return (
                 <View
                   key={index}
